+++ Unity/Runtime/Menu/IMenuOptionUI.cs
@@ -1,0 +1,7 @@
+namespace Seed.Unity.Menu
+{
+    public interface IMenuOptionUI
+    {
+        void SetText(string text);
+    }
+}